+++ io-core/src/nativeMain/kotlin/io/github/fukusaka/keel/io/DefaultAllocator.native.kt
@@ -1,0 +1,3 @@
+package io.github.fukusaka.keel.io
+
+actual fun defaultAllocator(): BufferAllocator = SlabAllocator()